Move to Illuminate DAL
* Moved to Laravel's database abstraction layer.
* Rewrote all queries, inserts and updates
* The whole DB schema is now versioned with schema changes

+++ src/Modules/TRICKLE_MODULE/Trickle.php
@@ -15,5 +15,5 @@ class Trickle extends DBRow {
 	public float $amountSta;
 	public float $amountStr;
 	public float $amountSen;
-	public ?int $amount;
+	public ?float $amount;
 }
